### BFF unification: one shared bearer/proxy/session seam (2026-06-27)
Extracted the per-app BFF plumbing (previously triplicated and divergent across
portal/base/landing) into `@aomi-labs/account` (server-only) so every Next app
mounts it as a thin config — change the auth model once, all clients inherit it.
