+++ apps/portal/src/server/cookies/session.ts
@@ -1,91 +1,16 @@
 import "server-only";
 
-import type { NextRequest, NextResponse } from "next/server";
-import { jwtVerify, SignJWT } from "jose";
-
 /**
- * The portal's **own** session — a stripped-down, Better-Auth-style session
- * cookie that carries our canonical user id. This is the identity root: it is
- * established by *us* and is independent of any provider. Privy/Para sign-in is
- * a separate, downstream act that *links a credential* and resolves which
- * canonical user this session belongs to; once the cookie is set, the session
- * stands on its own — "we own our whole cookie creation, not dependent on
- * Privy" (service-identity.md, "Identity root").
- *
- * The cookie is an HS256 JWT (`sub` = canonical user id) signed with
- * `PORTAL_ONLY_SESSION_SECRET` — a **symmetric** HMAC key. The portal both signs
- * and verifies it, so for this cookie verify == forge: the secret must live ONLY
- * on the portal and is never shared with the backend (which never sees this
- * cookie — the proxy strips it). httpOnly so the browser never reads it. The
- * AccountBearer for the backend is *derived from this session* (see
- * `mintAccountBearer`), not carried by the browser — the proxy injects it
- * server-side from this cookie (Option 2). This is the in-repo precursor to
- * arixoneth's Better-Auth session; the cookie name + claim are the seam.
+ * The portal session cookie now lives in `@aomi-labs/account` so every Aomi BFF
+ * (portal, base, landing) shares one session shape and one signing convention.
+ * This file is a thin re-export kept so existing `@portal/server/cookies/session`
+ * imports stay valid — import from `@aomi-labs/account` directly in new code.
  */
-export const SESSION_COOKIE = "aomi_session";
-const SESSION_TTL_SECONDS = 7 * 24 * 60 * 60;
-
-function secret(): Uint8Array {
-  const value = process.env.PORTAL_ONLY_SESSION_SECRET?.trim();
-  if (!value || value.length < 16) {
-    throw new Error(
-      "PORTAL_ONLY_SESSION_SECRET is not set (or too short) — the portal session cookie needs a signing secret",
-    );
-  }
-  return new TextEncoder().encode(value);
-}
-
-/** Sign a session cookie binding the browser to a canonical user id. */
-export async function issueSessionCookie(canonicalUserId: string): Promise<string> {
-  const now = Math.floor(Date.now() / 1000);
-  return new SignJWT({})
-    .setProtectedHeader({ alg: "HS256" })
-    .setSubject(canonicalUserId)
-    .setIssuedAt(now)
-    .setExpirationTime(now + SESSION_TTL_SECONDS)
-    .sign(secret());
-}
-
-/** Verify a session cookie and return its canonical user id, or null. */
-export async function readSessionCookie(
-  token: string | undefined,
-): Promise<string | null> {
-  if (!token) return null;
-  try {
-    const { payload } = await jwtVerify(token, secret());
-    return typeof payload.sub === "string" && payload.sub ? payload.sub : null;
-  } catch {
-    return null;
-  }
-}
-
-/** Set the session cookie on a response (called after login resolves the user). */
-export async function setSessionCookie(
-  response: NextResponse,
-  canonicalUserId: string,
-): Promise<void> {
-  const token = await issueSessionCookie(canonicalUserId);
-  response.cookies.set(SESSION_COOKIE, token, {
-    httpOnly: true,
-    secure: process.env.NODE_ENV === "production",
-    sameSite: "lax",
-    path: "/",
-    maxAge: SESSION_TTL_SECONDS,
-  });
-}
-
-/** Read the canonical user id bound to the request's session cookie, or null. */
-export async function getSessionedCanonicalId(request: NextRequest): Promise<string | null> {
-  return readSessionCookie(request.cookies.get(SESSION_COOKIE)?.value);
-}
-
-/** Clear the session cookie (sign-out). */
-export function clearSessionCookie(response: NextResponse): void {
-  response.cookies.set(SESSION_COOKIE, "", {
-    httpOnly: true,
-    secure: process.env.NODE_ENV === "production",
-    sameSite: "lax",
-    path: "/",
-    maxAge: 0,
-  });
-}
+export {
+  SESSION_COOKIE,
+  issueSessionCookie,
+  readSessionCookie,
+  setSessionCookie,
+  getSessionedCanonicalId,
+  clearSessionCookie,
+} from "@aomi-labs/account";
