+++ packages/account/package.json
@@ -17,7 +17,11 @@
   },
   "dependencies": {
     "@aomi-labs/service": "workspace:*",
+    "jose": "^6.1.3",
     "pg": "^8.22.0"
+  },
+  "peerDependencies": {
+    "next": ">=15"
   },
   "devDependencies": {
     "@types/pg": "^8.20.0"
