+++ packages/account/src/index.ts
@@ -1,8 +1,9 @@
-// `@aomi-labs/account` — server-only account graph for the issuer (the portal
-// BFF). Resolve-or-create the canonical user against the database the Rust
-// backend reads, and mint the AccountBearer that carries `sub` = that canonical
-// id. Node-only (holds `pg` + the EdDSA private key) — never import into a
-// browser/client bundle. See
+// `@aomi-labs/account` — server-only BFF seam for every Aomi frontend (portal,
+// base, landing). Resolve-or-create the canonical user against the database the
+// Rust backend reads, mint the AccountBearer that carries `sub` = that canonical
+// id, and provide the shared session cookie + backend proxy + auth-exchange
+// route every BFF mounts. Node-only (holds `pg` + the EdDSA private key) — never
+// import into a browser/client bundle. See
 // docs/topics/account-authentication/facts/service-identity.md.
 
 export {
@@ -17,3 +18,21 @@ export {
   type MintedBearer,
 } from "./bearer";
 export { portalService } from "./topology";
+export {
+  SESSION_COOKIE,
+  issueSessionCookie,
+  readSessionCookie,
+  setSessionCookie,
+  getSessionedCanonicalId,
+  clearSessionCookie,
+} from "./session";
+export {
+  createBackendProxy,
+  type ProxyConfig,
+  type AllowedRoute,
+} from "./proxy";
+export {
+  createAuthExchangeRoute,
+  type ExchangeConfig,
+  type Provider,
+} from "./exchange";
